// src/app/model/musicpiece.ts
import {language} from "./language";

export class musicpiece{
  id: number;
  title: string;
  artist: string
  language: language;
  theme: string;
  genre: string;
  instrumentType: string;
  text: string;
  youtubeURL: string;
  difficulty: string;
  typeofpiece: string;
  rating: number;
  myRating: number;
  musicxml: string|any;
  audio: any;

}